Remove the obsolete mno_webview package
The mno_webview package was removed from the project due to it not being needed anymore. The specific code related to it was also removed or adjusted accordingly in the other files in the project. The dependency on flutter_inappwebview was added directly to the necessary parts of the application.